Impl POC API to add/rm/modify CookieStore
* changes the signature for `ClientBuilder::cookie_store` to take a
`CookieStore` to use, instead of a `bool` to initialize a default
`CookieStore`
* add `Client::take_cookie_store` to retrieve a `CookieStore` from a
`Client`
* add `Client::replace_cookie_store` to replace a `CookieStore`
currently contained in a `Client`, returning any prior `CookieStore`
* add `Client::modify_cookie_store` to obtain mutable access to the
current `CookieStore`.

* Inverts the `Option`/`RwLock` nesting of `ClientRef::cookie_store`.
This is necessary to be able to safely remove or replace a `CookieStore`
at some point, but means we must acquire the lock in all cases (even
when a `CookieStore` is not utilized).
* Exposes the `cookie_store::CookieStore` struct in `reqwest`s public
API