Implement cookie storage via `cookie::Session`
Introduces a wrapper type `Session` which holds an implementer
of the trait `CookieStorage` along with a `Client`. Allows the
user to pass in the initial cookie storage, run a series of
requests and potentially modify the store during the session,
and then finally `end()` the `Session` to retrieve the resulting
cookie storage.