Update README advice regarding OpenSSL
Previously, the README suggested that only certain older versions of
OpenSSL are supported. This is no longer accurate (e.g. OpenSSL 3 works),
the better source of truth for supported versions is https://docs.rs/openssl.

Also adds clarity that OpenSSL can be vendored, useful if not installed
on the host.