crate: hide internal features (#2026)
In `Cargo.toml`, hide all optional dependencies as features

Currently, reqwest publishes all optional dependencies as features - which is usually not what is intended. Instead, the features should be explicitly declared, whereas optional features are enabled with the new `dep:*` syntax.

Note that I also had to fix a few conditional compilations which used optional crates rather than features as conditionals.

Also, note the use of the `...?/...` syntax instead of `.../...` -- as this both enabled an optional crate AND adds a feature to it.  Now, it uses the `"dep:...", "...?/..."` syntax to prevent exposing crate name.

Technically, this is *breaking change*, since people may have enabled some of these features accidentally. That's why we're including it in 0.12.