tests: bypass the proxy when testing timeouts
If an explicit proxy is configured in the environment, the timeout tests
will actually go through it, and then the test results will depend on
the proxy's performance, which can vary.

Testing this is very simple:
```
export http_proxy="http://127.0.0.1:1234"
cargo test timeouts
```

We're hitting this issue in the Ubuntu CI. Further investigations
after 892569e10b69d1a0e5db1f53202d1ebf09924fc1 seemed to point that
**sometimes**, the proxy would be too slow (particularly on non amd64
architectures), leading to the tests to actually timeout and pass.