Enable happy eyeballs when using `hickory-dns`
Happy Eyeballs algorithm is implemented for hyper, however it is not
working correctly for IPv6 only hosts because the default resolver
option in hickory is `Ipv4ThenIpv6`, meaning it only sends AAAA queries
if it cannot resolve an IPv4 address. Thus the address list `hyper`
receives is almost always IPv4 only given most servers have an IPv4
address. To make the Happy Eyeballs algorithm work correctly, we need
the resolver to resolve both IP versions. This also aligns with the
default GAI resolver behavior for both glibc and musl.