tests: use a documented test network for testing
10.255.255.0/24 is part of 10.0.0.0/8 and reserved for "Private Use" by
IANA [1]. That means that those IP might be in use in some internal
networks, making the timeout tests to fail on those networks.
One example of such network is the Ubuntu testing infrastructure.

Changing for a documented "Test network", that is supposed to never be
routable [2] makes the tests to have a better chance to pass in most
network contexts.

[1]: https://www.iana.org/assignments/iana-ipv4-special-registry/iana-ipv4-special-registry.xhtml
[2]: https://www.rfc-editor.org/rfc/rfc5737.html