Re-enable ability to set NoProxy via env on windows
Closes #2599

In #2559, the behavior of no proxy handling on windows was changed to
effectively only read from the current user's `Internet Settings` in the
registry on Windows. This meant that NO_PROXY env vars were no longer
respected on Windows. This commit changes that behavior and at least
brings it in line with dotnet and nuget behavior (two microsoft
projects) of first checking env vars and only if they are not set, check
the registry settings.