Fix the behavior of `NoProxy::from_env` to match docs
This fix changes the behavior of `NoProxy::from_env` if `NO_PROXY` or
`no_proxy` is set to an empty string. Previously, it would have returned
`None`, which is not what the documentation says should happen. It now
returns an `Some(NoProxy::default())` which should functionally be the
same thing for the purposes of not matching anything with the proxy.

This change was done due to potential interaction with previous commit
fixing #2599. I would expect that setting `NO_PROXY` to an empty string
would also prevent reqwest from reading the registry settings.