Re-enable NO_PROXY envs on Windows (#2601)
In #2559, the behavior of no proxy handling on windows was changed to
effectively only read from the current user's `Internet Settings` in the
registry on Windows. This meant that NO_PROXY env vars were no longer
respected on Windows. This commit changes that behavior and at least
brings it in line with dotnet and nuget behavior (two microsoft
projects) of first checking env vars and only if they are not set, check
the registry settings.

This fix changes the behavior of `NoProxy::from_env` if `NO_PROXY` or
`no_proxy` is set to an empty string. Previously, it would have returned
`None`, which is not what the documentation says should happen. It now
returns an `Some(NoProxy::default())` which should functionally be the
same thing for the purposes of not matching anything with the proxy.

This change was done due to potential interaction with previous commit
fixing #2599. I would expect that setting `NO_PROXY` to an empty string
would also prevent reqwest from reading the registry settings.

Closes #2599