Flip incorrect logic for dual backends
There was a piece of code in the client builder that has the following
code comment:

"Default backend + rustls Identity doesn't work."

It appears however that this piece of code was being run when rustls is
enabled, and native-tls is DISABLED, which seems to be a mistake. This
changes the feature flag to only run when both rustls and the native
backend are BOTH enabled.