proxy: allow setting dynamic headers at intercept time
To support this, I need the callback in `Proxy::custom` to be able to
return auth/headers alongside a `Url`.

I've packaged the auth/headers as `ProxyExtra` (formerly `Extra`), which
is now part of the public API.

Preserving backwards compat needed some minor trait contortions.
Technically there's a new `Sized` trait bound on `IntoProxy`, but this
_should_ be OK, since I'm not sure how you would impl the trait for an
unsized type anyway.